global: split cli.py into different modules
* `cli.py` was getting very big and unmanageable, around 2500 LOC,
  since we expect the tool to grow quickly we are splitting it
  into several modules.
