+++ setup.py
@@ -67,9 +67,9 @@ setup(
     author="REANA",
     author_email="info@reana.io",
     url="http://www.reana.io/",
-    packages=["reana"],
+    packages=packages,
     zip_safe=False,
-    entry_points={"console_scripts": ["reana-dev = reana.cli:cli",],},
+    entry_points={"console_scripts": ["reana-dev = reana.reana_dev.cli:reana_dev",],},
     extras_require=extras_require,
     install_requires=install_requires,
     setup_requires=setup_requires,
